Use ets instead of agent process.
Previous implementation relied on an Agent process that implied a
GenServer call on every log entry, disabled or not. This was particulary
bad on concurrent access, since every log entry have to wait for
previous requests to the Agent before doing the proper logging.

Current implementation uses an ETS table to read if tags are enabled or
not on each log level, and this avoid expensive and serialized calls to
another process. The GenServer holding the ETS table will be called only
once per tag if such tag is not yet into the table.

Previously Logbook was ~2% slower than Logger for enabled log levels,
and ~52% slower for disabled log levels (that basically outputs
nothing).

Now for enabled log levels ~1.2% slower and for disabled levels is only ~3%
slower.